Move thread wrapper to utils module
We're going to need to re-use this in multiple places.
